Complete 02 margin and padding
text-align: center for the div and margin: 0 auto for the button to
center both. The rest again was merely filling in the pixel values for
the margin and padding, except the font-size of the title also had to
be changed. Try to keep the margin overlapping simpler and remove
unnecessary ones.